requirements.txt
django>=1.11,<2.0
pytz
pymongo>=3.7.2,<3.8.0
ipython>=7.4.0,<7.5.0
python-twitter>=3.4.2,<3.5.0
python-decouple>=3.1,<3.2
